requirements.txt
asgiref==3.5.2
autopep8==1.6.0
crispy-bootstrap5==0.6
dj-database-url==2.0.0
Django==3.2.19
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-on-heroku==1.1.2
flake8==4.0.1
fontawesome-free==5.15.3
fontawesomefree==6.1.1
gunicorn==20.1.0
importlib-metadata==4.2.0
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
Pillow==9.2.0
pio==0.0.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2023.3
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
whitenoise==6.2.0
zipp==3.15.0
